+++ db/feed.go
@@ -12,11 +12,11 @@ type Video struct {
 	gorm.Model
 	UpdatedAt     time.Time `gorm:"column:update_time;not null;index:idx_update" `
 	Author        User      `gorm:"foreignkey:AuthorID"`
-	AuthorID      int       `gorm:"index:idx_authorid;not null"`
+	AuthorID      int64     `gorm:"index:idx_authorid;not null"`
 	PlayUrl       string    `gorm:"type:varchar(255);not null"`
 	CoverUrl      string    `gorm:"type:varchar(255)"`
-	FavoriteCount int       `gorm:"default:0"`
-	CommentCount  int       `gorm:"default:0"`
+	FavoriteCount int64     `gorm:"default:0"`
+	CommentCount  int64     `gorm:"default:0"`
 	Title         string    `gorm:"type:varchar(50);not null"`
 }
